Revert "Merge remote-tracking branch 'origin/swap-improve-balance-tests' into swap-segregate-api"
This reverts commit 456e7c6851de925f4f39c63ee0f2a382362bad2d, reversing
changes made to 4177f1d9d70fcb203ad008387c3b530c3698100a.
Revert "swap: move consts into swap/api.go, move api tests into api_test.go"

This reverts commit 4177f1d9d70fcb203ad008387c3b530c3698100a.
Revert "Merge remote-tracking branch 'origin/swap-improve-balance-tests' into swap-segregate-api"

This reverts commit 95e4970504cbabd514e5e8888e60b475fcdc386f, reversing
changes made to cd6c4965a79e86200b3dd068082890bee3d8d46a.
